Consolidate listener registration
