+++ src/Agent.Core/DiagnosticSource/ElasticCoreListeners.cs
@@ -1,0 +1,25 @@
+﻿using System;
+using System.Collections.Generic;
+using Elastic.Agent.Core.DiagnosticListeners;
+
+namespace Elastic.Agent.Core.DiagnosticSource
+{
+    /// <summary>
+    /// Manages all listeners that are generated by Types which are part of .netstandard 2.0
+    /// </summary>
+    public class ElasticCoreListeners
+    {
+        private Config _agentConfig = new Config(); //TODO: Config should be passed from outside
+
+        /// <summary>
+        /// Start listening for diagnosticsource events. Only listens for sources that are part of the Agent.Core package.
+        /// </summary>
+        public void Start()
+        {
+            System.Diagnostics.DiagnosticListener
+            .AllListeners
+            .Subscribe(new DiagnosticInitializer(new List<IDiagnosticListener> {new HttpDiagnosticListener(_agentConfig) }));
+
+        }
+    }
+}
